Code sharing is good.
